gc_ctrl.c: gros bug dans Gc.stat instrtrace.c: warning gcc sur printf memory.h: assertion ajoutee
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@1565 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/instrtrace.c
@@ -34,7 +34,7 @@ void disasm_instr(pc)
      code_t pc;
 {
   int instr = *pc;
-  printf("%6d  %s", pc - start_code,
+  printf("%6ld  %s", (long) (pc - start_code),
          instr < 0 || instr > STOP ? "???" : names_of_instructions[instr]);
   pc++;
   switch(instr) {
